Changed deployment target and fixed argument type in PPCameraOverlayViewController cameraViewController:didPublishProgress:

+++ Samples/pdf417-sample/pdf417-sample/Camera Overlays/Simple/PPCameraOverlayViewController.m
@@ -108,7 +108,7 @@
     NSLog(@"Barcode scanning process terminated");
 }
 
-- (void)cameraViewController:(id<PPScanningViewController>)cameraViewController didPublishProgress:(float)progress {
+- (void)cameraViewController:(id<PPScanningViewController>)cameraViewController didPublishProgress:(CGFloat)progress {
     NSLog(@"Barcode scanning don't have progress reporting implemented yet");
 }
